debs: Remove usage of distro variants
There is a fair amount of symlinking and reference to Debian/Ubuntu
versions, which is not required (and probably hasn't been for a while).
When building and publishing debs, we publish them to a single apt repo
(https://packages.cloud.google.com/apt/dists/kubernetes-xenial) which
makes no distinction between the packages within.

This commit moves the deb package definitions to `packages/latest` and
removes all symlinks. The deb build program is also slightly refactored
here to reference the true distribution types (stable, testing, unstable).

Future commits will divide the package definitions so we can modify the
structure and dependencies of packages for specific versions of
Kubernetes without affecting previous versions.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>